+++ .github/workflows/build.yml
@@ -92,11 +92,6 @@ jobs:
           just npm prune
           cd warpgate-web && NODE_ENV=dev cyclonedx-npm --output-format xml > ../cdx/admin-ui.cdx.xml
 
-      - name: Build RDP helper
-        run: just build-rdp-helper --target ${{ matrix.target }}
-        env:
-          SOURCE_DATE_EPOCH: "0" # for embedded-helper determinism
-
       - name: Build
         uses: actions-rs/cargo@844f36862e911db73fe0815f00a4a2602c279505
         with:
