+++ .github/workflows/clippy.yml
@@ -20,9 +20,6 @@ jobs:
           cargo install --locked cargo-cranky@0.3.0
 
       # Built first so warpgate-protocol-rdp's embedded-helper code paths are linted.
-      - name: Build RDP helper
-        run: just build-rdp-helper
-
       # warpgate-web embeds ./dist via RustEmbed, which needs the folder to exist
       # at compile time. The lint job doesn't ship assets, so an empty dir is enough.
       - name: Stub web assets for RustEmbed
