+++ .github/workflows/reprotest.yml
@@ -30,4 +30,4 @@ jobs:
       - name: Reprotest
         run: |
           sudo ulimit -n 999999
-          sudo env "PATH=$PATH" reprotest -vv --min-cpus=99999 --vary=environment,build_path,kernel,aslr,num_cpus,-time,-user_group,fileordering,domain_host,home,locales,exec_path,timezone,umask --build-command 'just npm ci; just npm run build; SOURCE_DATE_EPOCH=0 just build-rdp-helper; SOURCE_DATE_EPOCH=0 cargo build --all-features --release' . target/release/warpgate
+          sudo env "PATH=$PATH" reprotest -vv --min-cpus=99999 --vary=environment,build_path,kernel,aslr,num_cpus,-time,-user_group,fileordering,domain_host,home,locales,exec_path,timezone,umask --build-command 'just npm ci; just npm run build; SOURCE_DATE_EPOCH=0 cargo build --all-features --release' . target/release/warpgate
