+++ justfile
@@ -1,14 +1,10 @@
-projects := "warpgate warpgate-admin warpgate-common warpgate-db-entities warpgate-db-migrations warpgate-database-protocols warpgate-protocol-ssh warpgate-protocol-mysql warpgate-protocol-postgres warpgate-protocol-kubernetes warpgate-protocol-http warpgate-protocol-rdp warpgate-rdp-helper warpgate-protocol-vnc warpgate-core warpgate-sso"
+projects := "warpgate warpgate-admin warpgate-common warpgate-db-entities warpgate-db-migrations warpgate-database-protocols warpgate-protocol-ssh warpgate-protocol-mysql warpgate-protocol-postgres warpgate-protocol-kubernetes warpgate-protocol-http warpgate-protocol-rdp warpgate-protocol-vnc warpgate-core warpgate-sso"
 
 run *ARGS='run':
     RUST_BACKTRACE=1 cargo run --all-features -- --config config.yaml {{ARGS}}
 
 run-release *ARGS='run':
     RUST_BACKTRACE=1 cargo run --all-features --release -- --config config.yaml {{ARGS}}
-
-# Run before `cargo build` so warpgate-protocol-rdp's build script can embed it into the main binary.
-build-rdp-helper *ARGS:
-    cargo build --release -p warpgate-rdp-helper {{ARGS}}
 
 fmt:
     for p in {{projects}}; do cargo fmt -p $p -v; done
