+++ warpgate-protocol-rdp/src/client/tls.rs
@@ -1,0 +1,118 @@
+//! TLS setup for the target-facing RDP connection.
+
+use std::sync::Arc;
+
+use anyhow::{Context, Result};
+use ironrdp_server::tokio_rustls::TlsConnector;
+use ironrdp_server::tokio_rustls::client::TlsStream;
+use tokio::net::TcpStream;
+
+/// Wrap `stream` in TLS and return it alongside the server's public key, which CredSSP
+/// channel-binds to.
+pub async fn upgrade(
+    stream: TcpStream,
+    server_name: String,
+    verify: bool,
+) -> Result<(TlsStream<TcpStream>, Vec<u8>)> {
+    let mut config = if verify {
+        let mut roots = rustls::RootCertStore::empty();
+        for cert in rustls_native_certs::load_native_certs().certs {
+            roots.add(cert).ok();
+        }
+        rustls::client::ClientConfig::builder()
+            .with_root_certificates(roots)
+            .with_no_client_auth()
+    } else {
+        // RDP servers commonly use self-signed certs, so verification is off by default;
+        // CredSSP/NLA still channel-binds to the server's public key.
+        rustls::client::ClientConfig::builder()
+            .dangerous()
+            .with_custom_certificate_verifier(Arc::new(danger::NoCertificateVerification))
+            .with_no_client_auth()
+    };
+    config.resumption = rustls::client::Resumption::disabled();
+
+    let connector = TlsConnector::from(Arc::new(config));
+    let server_name = server_name.try_into().context("invalid server name")?;
+    let tls_stream = connector
+        .connect(server_name, stream)
+        .await
+        .context("TLS handshake")?;
+
+    let cert = tls_stream
+        .get_ref()
+        .1
+        .peer_certificates()
+        .and_then(|certs| certs.first())
+        .context("missing peer certificate")?;
+    let server_public_key = extract_server_public_key(cert)?;
+
+    Ok((tls_stream, server_public_key))
+}
+
+fn extract_server_public_key(cert: &[u8]) -> Result<Vec<u8>> {
+    use x509_cert::der::Decode as _;
+    let cert = x509_cert::Certificate::from_der(cert).context("parsing certificate")?;
+    let key = cert
+        .tbs_certificate
+        .subject_public_key_info
+        .subject_public_key
+        .as_bytes()
+        .context("public key not byte-aligned")?
+        .to_owned();
+    Ok(key)
+}
+
+mod danger {
+    use rustls::client::danger::{HandshakeSignatureValid, ServerCertVerified, ServerCertVerifier};
+    use rustls::pki_types::{CertificateDer, ServerName, UnixTime};
+    use rustls::{DigitallySignedStruct, Error, SignatureScheme};
+
+    #[derive(Debug)]
+    pub struct NoCertificateVerification;
+
+    impl ServerCertVerifier for NoCertificateVerification {
+        fn verify_server_cert(
+            &self,
+            _end_entity: &CertificateDer<'_>,
+            _intermediates: &[CertificateDer<'_>],
+            _server_name: &ServerName<'_>,
+            _ocsp: &[u8],
+            _now: UnixTime,
+        ) -> Result<ServerCertVerified, Error> {
+            Ok(ServerCertVerified::assertion())
+        }
+
+        fn verify_tls12_signature(
+            &self,
+            _message: &[u8],
+            _cert: &CertificateDer<'_>,
+            _dss: &DigitallySignedStruct,
+        ) -> Result<HandshakeSignatureValid, Error> {
+            Ok(HandshakeSignatureValid::assertion())
+        }
+
+        fn verify_tls13_signature(
+            &self,
+            _message: &[u8],
+            _cert: &CertificateDer<'_>,
+            _dss: &DigitallySignedStruct,
+        ) -> Result<HandshakeSignatureValid, Error> {
+            Ok(HandshakeSignatureValid::assertion())
+        }
+
+        fn supported_verify_schemes(&self) -> Vec<SignatureScheme> {
+            vec![
+                SignatureScheme::RSA_PKCS1_SHA256,
+                SignatureScheme::RSA_PKCS1_SHA384,
+                SignatureScheme::RSA_PKCS1_SHA512,
+                SignatureScheme::ECDSA_NISTP256_SHA256,
+                SignatureScheme::ECDSA_NISTP384_SHA384,
+                SignatureScheme::RSA_PSS_SHA256,
+                SignatureScheme::RSA_PSS_SHA384,
+                SignatureScheme::RSA_PSS_SHA512,
+                SignatureScheme::ED25519,
+            ]
+        }
+    }
+}
