+++ warpgate-protocol-rdp/src/lib.rs
@@ -1,47 +1,33 @@
 //! In-workspace RDP integration for Warpgate.
 //!
-//! The actual RDP/IronRDP work runs in the standalone `warpgate-rdp-helper` binary
-//! (which lives outside the cargo workspace to avoid a RustCrypto pre-release version
-//! conflict between IronRDP's CredSSP stack and `russh`). The prebuilt helper is
-//! embedded into this crate at build time (see `build.rs`) and extracted for
-//! use, so Warpgate ships as a single executable. This crate spawns that helper as a
-//! subprocess and bridges its length-delimited binary stdio (raw BGRA frames,
-//! JSON control messages) to the shared [`DesktopEvent`]/[`DesktopInput`] streams,
-//! so the existing web-desktop manager and browser canvas renderer work unchanged.
+//! [`client`] drives IronRDP against a target host; [`server`] runs IronRDP's server state
+//! machine for native RDP viewers (mstsc/FreeRDP) connecting to Warpgate's RDP port. Both
+//! speak the shared [`DesktopEvent`]/[`DesktopInput`] streams, so the web-desktop manager
+//! and browser canvas renderer work against either front end unchanged.
 
-mod embedded;
-mod helper;
+mod client;
 mod server;
 mod session_handle;
 
-use std::process::Stdio;
-
 use anyhow::Context;
-use bytes::Bytes;
 use futures::future::BoxFuture;
-use futures::{SinkExt, StreamExt};
 pub use server::bind_server;
-use tokio::io::{AsyncBufReadExt, BufReader};
-use tokio::sync::mpsc::{
-    Receiver, Sender, UnboundedReceiver, UnboundedSender, channel, unbounded_channel,
-};
-use tokio_util::codec::{FramedRead, FramedWrite, LengthDelimitedCodec};
-use tracing::{Instrument, debug, error, info_span, warn};
-use warpgate_common::{ListenEndpoint, ProtocolName, RdpTargetAuth, TargetRdpOptions};
+use tokio::sync::mpsc::{Receiver, Sender, UnboundedSender, channel, unbounded_channel};
+use tracing::{Instrument, error, info_span};
+use warpgate_common::{ListenEndpoint, ProtocolName, TargetRdpOptions};
 use warpgate_core::{
-    DESKTOP_INPUT_CHANNEL_CAPACITY, DesktopEvent, DesktopInput, DesktopRect, DesktopState,
-    ProtocolServer, Services,
+    DESKTOP_INPUT_CHANNEL_CAPACITY, DesktopEvent, DesktopInput, DesktopState, ProtocolServer,
+    Services,
 };
-use warpgate_rdp_ipc::client::{ConnectConfig, Event as HelperEvent, Input as HelperInput};
 use warpgate_tls::TlsCertificateAndPrivateKey;
 
 pub static PROTOCOL_NAME: ProtocolName = "RDP";
 
-pub use warpgate_rdp_ipc::DEFAULT_SIZE;
+pub use warpgate_desktop_ui::DEFAULT_SIZE;
 
 /// The native RDP server endpoint. Standard RDP clients (mstsc/FreeRDP) connect
 /// directly to Warpgate's RDP port; per connection it brokers between the viewer-facing
-/// serve helper and the existing target-facing client helper (see [`server`]).
+/// RDP server and a target-facing client (see [`server`]).
 pub struct RdpProtocolServer {
     services: Services,
 }
@@ -61,7 +47,6 @@ impl ProtocolServer for RdpProtocolServer {
         proxy_protocol: bool,
         tls: Vec<TlsCertificateAndPrivateKey>,
     ) -> anyhow::Result<BoxFuture<'static, anyhow::Result<()>>> {
-        // The serve helper terminates TLS itself, so hand it the raw PEM.
         let certificate_and_key = tls
             .into_iter()
             .next()
@@ -84,14 +69,14 @@ impl std::fmt::Debug for RdpProtocolServer {
     }
 }
 
-/// Handles for driving a backend RDP client (running in the helper subprocess).
+/// Handles for driving a backend RDP client.
 pub struct RdpClientHandles {
     pub event_rx: Receiver<DesktopEvent>,
     pub input_tx: Sender<DesktopInput>,
     pub abort_tx: UnboundedSender<()>,
 }
 
-/// Spawn the RDP helper for a target and bridge it to normalised desktop streams.
+/// Start an RDP client for a target and bridge it to normalised desktop streams.
 pub fn connect(options: TargetRdpOptions, size: (u16, u16)) -> RdpClientHandles {
     let (event_tx, event_rx) = channel::<DesktopEvent>(1024);
     let (input_tx, input_rx) = channel::<DesktopInput>(DESKTOP_INPUT_CHANNEL_CAPACITY);
@@ -100,8 +85,10 @@ pub fn connect(options: TargetRdpOptions, size: (u16, u16)) -> RdpClientHandles 
     let span = info_span!("RDP-client", host = %options.host, port = options.port);
     tokio::spawn(
         async move {
-            if let Err(error) = run(options, size, event_tx.clone(), input_rx, abort_rx).await {
-                error!(%error, "RDP helper failed");
+            if let Err(error) =
+                client::run(options, size, event_tx.clone(), input_rx, abort_rx).await
+            {
+                error!(%error, "RDP client failed");
                 let _ = event_tx.send(DesktopEvent::Error(error.to_string())).await;
             }
             let _ = event_tx
@@ -116,180 +103,4 @@ pub fn connect(options: TargetRdpOptions, size: (u16, u16)) -> RdpClientHandles 
         input_tx,
         abort_tx,
     }
-}
-
-async fn run(
-    options: TargetRdpOptions,
-    (width, height): (u16, u16),
-    event_tx: tokio::sync::mpsc::Sender<DesktopEvent>,
-    mut input_rx: Receiver<DesktopInput>,
-    mut abort_rx: UnboundedReceiver<()>,
-) -> anyhow::Result<()> {
-    event_tx
-        .send(DesktopEvent::State(DesktopState::Connecting))
-        .await
-        .ok();
-
-    let password = match &options.auth {
-        RdpTargetAuth::Password(auth) => auth.password.expose_secret().clone(),
-    };
-
-    // Kept in scope until after `spawn` so the Linux memfd stays open across exec.
-    let helper = helper::resolve()?;
-
-    let mut child = tokio::process::Command::new(helper.path())
-        .arg("connect")
-        .stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped())
-        // Kill the helper if this task is cancelled/dropped (tokio doesn't by default).
-        .kill_on_drop(true)
-        .spawn()
-        .with_context(|| format!("spawning RDP helper ({})", helper.path().display()))?;
-
-    let stdin = child.stdin.take().context("helper stdin")?;
-    let stdout = child.stdout.take().context("helper stdout")?;
-
-    // Surface helper diagnostics (panics, errors) to the log instead of discarding them.
-    if let Some(stderr) = child.stderr.take() {
-        tokio::spawn(async move {
-            let mut lines = BufReader::new(stderr).lines();
-            while let Ok(Some(line)) = lines.next_line().await {
-                debug!(helper = %line, "RDP helper stderr");
-            }
-        });
-    }
-
-    // Length-delimited framing (matches the helper). Frames can be a full-screen rect.
-    let codec = || {
-        LengthDelimitedCodec::builder()
-            .max_frame_length(warpgate_rdp_ipc::MAX_FRAME_LEN)
-            .new_codec()
-    };
-
-    // First frame: the connection config.
-    let mut stdin_wr = FramedWrite::new(stdin, codec());
-    let config = ConnectConfig {
-        host: options.host.clone(),
-        port: options.port,
-        username: options.username.clone(),
-        password,
-        domain: options.domain.clone(),
-        width,
-        height,
-        verify_tls: options.verify_tls,
-    };
-    let mut config_buf = Vec::new();
-    warpgate_rdp_ipc::encode_json_into(&config, &mut config_buf);
-    stdin_wr.send(Bytes::copy_from_slice(&config_buf)).await?;
-
-    // Forward input to the helper.
-    let input_task = tokio::spawn(async move {
-        let mut buf = Vec::new();
-        while let Some(input) = input_rx.recv().await {
-            let msg = match input {
-                DesktopInput::Pointer { x, y, buttons } => {
-                    Some(HelperInput::Pointer { x, y, buttons })
-                }
-                DesktopInput::Key { keysym, down } => Some(HelperInput::Key { keysym, down }),
-                DesktopInput::Scancode {
-                    code,
-                    extended,
-                    down,
-                } => Some(HelperInput::Scancode {
-                    code,
-                    extended,
-                    down,
-                }),
-                DesktopInput::Wheel {
-                    vertical, delta, ..
-                } => Some(HelperInput::Wheel {
-                    vertical,
-                    // RDP wheel rotation units are ~120 per notch.
-                    delta: delta.saturating_mul(120),
-                }),
-                // Clipboard/refresh not yet wired through the helper.
-                DesktopInput::Clipboard(_) | DesktopInput::Refresh => None,
-            };
-            if let Some(msg) = msg {
-                msg.encode_into(&mut buf);
-                if stdin_wr.send(Bytes::copy_from_slice(&buf)).await.is_err() {
-                    break;
-                }
-            }
-        }
-    });
-
-    // Read events from the helper.
-    let mut reader = FramedRead::new(stdout, codec());
-    loop {
-        tokio::select! {
-            biased;
-            _ = abort_rx.recv() => {
-                debug!("RDP client aborted");
-                break;
-            }
-            frame = reader.next() => {
-                let Some(frame) = frame else {
-                    break;
-                };
-                let frame = frame.context("reading helper output")?;
-                // `freeze()` is zero-copy; the image payload becomes a slice of this buffer.
-                if let Some(event) = HelperEvent::decode(&frame.freeze()) {
-                    // Race the (possibly blocking) send against abort so a slow consumer
-                    // can't starve abort handling while the helper floods stdout.
-                    tokio::select! {
-                        biased;
-                        _ = abort_rx.recv() => break,
-                        result = forward_event(&event_tx, event) => {
-                            if result.is_err() {
-                                break;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    input_task.abort();
-    let _ = child.kill().await;
-    Ok(())
-}
-
-async fn forward_event(
-    event_tx: &tokio::sync::mpsc::Sender<DesktopEvent>,
-    event: HelperEvent,
-) -> Result<(), ()> {
-    let mapped = match event {
-        HelperEvent::Connected { width, height } => {
-            event_tx
-                .send(DesktopEvent::State(DesktopState::Connected))
-                .await
-                .map_err(|_| ())?;
-            DesktopEvent::Resize { width, height }
-        }
-        HelperEvent::RawImage {
-            x,
-            y,
-            width,
-            height,
-            data,
-        } => DesktopEvent::RawImage {
-            rect: DesktopRect {
-                x,
-                y,
-                width,
-                height,
-            },
-            // `data` is a zero-copy `Bytes` slice of the wire frame — moved, not copied.
-            data,
-        },
-        HelperEvent::Error { message } => {
-            warn!(%message, "RDP backend error");
-            DesktopEvent::Error(message)
-        }
-        HelperEvent::Disconnected => DesktopEvent::State(DesktopState::Disconnected),
-    };
-    event_tx.send(mapped).await.map_err(|_| ())
 }
