+++ warpgate-protocol-rdp/src/server/mod.rs
@@ -1,72 +1,59 @@
 //! Native RDP server endpoint.
 //!
-//! Warpgate accepts a raw TCP connection from a standard RDP client (mstsc/FreeRDP)
-//! and brokers between two subprocesses of the out-of-workspace RDP helper:
+//! Warpgate accepts a raw TCP connection from a standard RDP client (mstsc/FreeRDP) and
+//! brokers between two halves:
 //!
-//! * the viewer-facing **serve helper** (`warpgate-rdp-helper serve`) runs the RDP server
-//!   state machine and terminates TLS, and
-//! * the target-facing **client helper** (`warpgate-rdp-helper connect`, via
-//!   [`crate::connect`]) drives the RDP client toward the configured target.
+//! * the viewer-facing [`rdp`] server runs the RDP server state machine and terminates
+//!   TLS, and
+//! * the target-facing [`crate::client`] (via [`crate::connect`]) drives the RDP client
+//!   toward the configured target.
 //!
 //! Warpgate keeps ownership of the listener, the session/audit lifecycle, and
-//! authentication. Per connection it creates a private, unnamed socketpair (the serve
-//! helper inherits one end as an fd), spawns the serve helper, and shuttles the raw (TLS)
-//! RDP byte stream between the viewer and the helper with [`copy_bidirectional`]. Credentials the viewer submits arrive over the helper's
-//! stdout control channel as [`ServerHelperEvent::AuthRequest`]; Warpgate authenticates
-//! them with the same [`AuthSelector`] flow used by the native VNC server, connects to
-//! the resolved target, and bridges target framebuffer events to the serve helper while
-//! recording them — so native RDP records exactly like the in-browser path.
+//! authentication. Credentials the viewer submits arrive as [`ServerEvent::AuthRequest`];
+//! Warpgate authenticates them with the same [`AuthSelector`] flow used by the native VNC
+//! server, connects to the resolved target, and bridges target framebuffer events back to
+//! the viewer while recording them — so native RDP records exactly like the in-browser
+//! path.
 
 use std::net::SocketAddr;
-use std::process::Stdio;
 use std::sync::Arc;
 
-use anyhow::{Context, Result};
+use anyhow::Result;
+use futures::FutureExt;
 use futures::future::BoxFuture;
-use futures::{FutureExt, SinkExt};
-use tokio::io::{AsyncBufReadExt, BufReader, copy_bidirectional};
+use tokio::io::copy_bidirectional;
 use tokio::net::TcpStream;
-use tokio::process::{ChildStdin, ChildStdout, Command};
-use tokio::sync::mpsc::{Sender, UnboundedSender, unbounded_channel};
+use tokio::sync::mpsc::{Sender, UnboundedReceiver, UnboundedSender, channel, unbounded_channel};
 use tokio_stream::StreamExt;
-use tokio_util::codec::{FramedRead, FramedWrite, LengthDelimitedCodec};
 use tracing::{Instrument, debug, error, info, info_span, warn};
 use warpgate_common::helpers::net::detect_port_knock;
 use warpgate_common::{ListenEndpoint, Target, TargetOptions, TargetRdpOptions};
 use warpgate_core::recordings::DesktopRecorder;
 use warpgate_core::{DesktopInput, Services, SessionStateInit, State, WarpgateServerHandle};
-use warpgate_rdp_ipc::server::{
-    Event as ServerHelperEvent, Input as ServerHelperInput, ServeConfig,
-};
-use warpgate_rdp_ipc::{DEFAULT_SCREEN_H, DEFAULT_SCREEN_W};
+use warpgate_desktop_ui::{DEFAULT_SCREEN_H, DEFAULT_SCREEN_W};
 
 use crate::PROTOCOL_NAME;
 use crate::session_handle::RdpSessionHandle;
 
 mod bridge;
 mod hold_screen;
-mod transport;
+mod protocol;
+mod rdp;
 
-use bridge::{connect_backend, helper_stdin_writer};
+use bridge::connect_backend;
 use hold_screen::run_hold_screen;
+use protocol::{Event as ServerEvent, Input as ServerInput};
 use warpgate_desktop_auth::{
     DesktopAuthOutcome, DesktopProtocol, authenticate, finalize_user_auth,
 };
 
-/// The fd number at which the serve helper receives its end of the RDP transport
-/// socketpair (`dup2`'d into place by `pre_exec`, then passed to the helper as a CLI arg).
-const HELPER_STREAM_FD: i32 = 3;
+/// How many pending messages the viewer-facing RDP server accepts before its bounded feed
+/// applies backpressure; `frame_bridge` sheds stale frames ahead of it.
+const SERVER_INPUT_CAPACITY: usize = 16;
 
-/// Length-delimited frame reader/writer over the serve helper's stdio. Frames can be a
-/// full-screen BGRA rect, so the size cap is raised past the codec's 8 MB default.
-type HelperReader = FramedRead<ChildStdout, LengthDelimitedCodec>;
-type HelperWriter = FramedWrite<ChildStdin, LengthDelimitedCodec>;
-
-fn helper_codec() -> LengthDelimitedCodec {
-    LengthDelimitedCodec::builder()
-        .max_frame_length(warpgate_rdp_ipc::MAX_FRAME_LEN)
-        .new_codec()
-}
+/// Size of each direction of the in-memory duplex carrying raw RDP bytes between the
+/// viewer socket and the RDP server.
+const RELAY_BUFFER_BYTES: usize = 64 * 1024;
 
 /// Handles to the connected target-side RDP client, kept once authentication succeeds.
 struct BackendBridge {
@@ -90,8 +77,6 @@ pub async fn bind_server(
     services: Services,
     address: ListenEndpoint,
     proxy_protocol: bool,
-    // The serve helper terminates TLS itself (it has its own rustls with a different
-    // crypto provider), so we hand it the raw PEM rather than a built acceptor.
     cert_pem: String,
     key_pem: String,
 ) -> Result<BoxFuture<'static, Result<()>>> {
@@ -178,106 +163,70 @@ async fn handle_connection(
         guard.wrap_stream(stream).await?
     };
 
-    // Warpgate owns the viewer socket + session; the serve helper runs the RDP server state
-    // machine. They're connected by a private, unnamed socketpair (see [`transport`]) — the
-    // helper inherits its end as fd HELPER_STREAM_FD, so there's no loopback port for another
-    // local process to connect to or race. Warpgate stays a transparent pipe of the raw (TLS)
-    // RDP bytes and never sees plaintext: TLS is terminated end-to-end between the viewer and
-    // the helper.
-    //
-    // Kept in scope until after `spawn` so the Linux memfd stays open across exec.
-    let helper = crate::helper::resolve()?;
-    let mut command = Command::new(helper.path());
-    command
-        .arg("serve")
-        .arg(HELPER_STREAM_FD.to_string())
-        .stdin(Stdio::piped())
-        .stdout(Stdio::piped())
-        .stderr(Stdio::piped())
-        // Kill the helper if this task is cancelled/dropped (tokio doesn't by default).
-        .kill_on_drop(true);
-    let (mut loopback, mut child) = transport::spawn_with_transport(&mut command)?;
+    // Warpgate owns the viewer socket and the session; `rdp` runs the RDP server state
+    // machine over it and terminates TLS toward the viewer.
+    let (server_out_tx, server_out_rx) = unbounded_channel::<ServerEvent>();
+    // Bounded so a slow viewer applies backpressure; `frame_bridge` sheds stale frames
+    // ahead of it, and control_loop's occasional auth verdicts share the same feed.
+    let (server_in_tx, server_in_rx) = channel::<ServerInput>(SERVER_INPUT_CAPACITY);
 
-    let child_stdin = child.stdin.take().context("serve helper stdin")?;
-    let child_stdout = child.stdout.take().context("serve helper stdout")?;
-
-    // Surface helper diagnostics (panics, errors) to the log instead of discarding them.
-    if let Some(stderr) = child.stderr.take() {
-        tokio::spawn(async move {
-            let mut lines = BufReader::new(stderr).lines();
-            while let Ok(Some(line)) = lines.next_line().await {
-                debug!(helper = %line, "RDP serve helper stderr");
-            }
-        });
-    }
-
-    // First frame on stdin: the serve config (TLS material + initial size).
-    let mut helper_stdin = FramedWrite::new(child_stdin, helper_codec());
-    let config = ServeConfig {
+    // `ironrdp-server` requires a `Sync` transport, which the session-wrapped viewer
+    // stream is not, so the server runs against an in-memory duplex and we relay the raw
+    // (TLS) RDP bytes across.
+    let (server_side, mut relay_side) = tokio::io::duplex(RELAY_BUFFER_BYTES);
+    let mut rdp_done = rdp::run_on_thread(
+        server_side,
         cert_pem,
         key_pem,
-        width: DEFAULT_SCREEN_W,
-        height: DEFAULT_SCREEN_H,
-    };
-    let mut config_buf = Vec::new();
-    warpgate_rdp_ipc::encode_json_into(&config, &mut config_buf);
-    helper_stdin
-        .send(bytes::Bytes::copy_from_slice(&config_buf))
-        .await?;
+        (DEFAULT_SCREEN_W, DEFAULT_SCREEN_H),
+        server_out_tx,
+        server_in_rx,
+    );
 
-    // All subsequent helper stdin writes are funnelled through one task.
-    let (helper_in_tx, helper_in_rx) = unbounded_channel::<ServerHelperInput>();
-    let stdin_task = tokio::spawn(helper_stdin_writer(helper_stdin, helper_in_rx));
-
-    // The socketpair is already connected — no connect-back to wait for. Run the raw byte
-    // relay (viewer ⇄ helper) and the JSON control loop concurrently;
-    // whichever ends first tears the session down.
+    // Run the RDP server and the control loop concurrently; whichever ends first tears
+    // the session down.
     tokio::select! {
-        result = control_loop(services, server_handle, remote_address, child_stdout, helper_in_tx) => {
+        result = control_loop(services, server_handle, remote_address, server_out_rx, server_in_tx) => {
             result?;
         }
-        result = copy_bidirectional(&mut viewer, &mut loopback) => {
+        result = &mut rdp_done => {
             match result {
-                Ok((to_helper, to_viewer)) => {
-                    debug!(to_helper, to_viewer, "RDP byte relay ended");
-                }
-                Err(error) => debug!(%error, "RDP byte relay error"),
+                Ok(Ok(())) => {}
+                Ok(Err(error)) => warn!(%error, "RDP server failed"),
+                Err(_) => debug!("RDP server ended"),
+            }
+        }
+        result = copy_bidirectional(&mut viewer, &mut relay_side) => {
+            if let Err(error) = result {
+                debug!(%error, "RDP byte relay error");
             }
         }
     }
 
-    stdin_task.abort();
-    let _ = child.kill().await;
     Ok(())
 }
 
-/// Read the serve helper's control channel: authenticate the viewer, connect to the
-/// target on success, and forward viewer input to the target client.
+/// Read the RDP server's event stream: authenticate the viewer, connect to the target on
+/// success, and forward viewer input to the target client.
 async fn control_loop(
     services: Services,
     server_handle: Arc<tokio::sync::Mutex<WarpgateServerHandle>>,
     remote_address: SocketAddr,
-    stdout: ChildStdout,
-    helper_in_tx: UnboundedSender<ServerHelperInput>,
+    mut events: UnboundedReceiver<ServerEvent>,
+    server_in_tx: Sender<ServerInput>,
 ) -> Result<()> {
-    let mut frames = FramedRead::new(stdout, helper_codec());
     let mut backend: Option<BackendBridge> = None;
-    // What the helper says the session settled on, which is what we paint and what we ask
-    // the target for. Seeded with the size we advertised, in case we dial before the
+    // What the RDP server says the session settled on, which is what we paint and what we
+    // ask the target for. Seeded with the size we advertised, in case we dial before the
     // capability exchange reports back.
     let mut screen = warpgate_desktop_ui::Screen {
         width: DEFAULT_SCREEN_W,
         height: DEFAULT_SCREEN_H,
     };
 
-    while let Some(frame) = frames.next().await {
-        let frame = frame.context("reading serve helper output")?;
-        let Some(event) = ServerHelperEvent::decode(&frame) else {
-            continue;
-        };
-
+    while let Some(event) = events.recv().await {
         let input = match event {
-            ServerHelperEvent::AuthRequest {
+            ServerEvent::AuthRequest {
                 username, password, ..
             } => {
                 if backend.is_some() {
@@ -302,7 +251,7 @@ async fn control_loop(
                             connect_backend(
                                 &services,
                                 &server_handle,
-                                &helper_in_tx,
+                                &server_in_tx,
                                 user_info,
                                 target,
                                 options,
@@ -310,8 +259,9 @@ async fn control_loop(
                             )
                             .await?,
                         );
-                        if helper_in_tx
-                            .send(ServerHelperInput::AuthResponse { accept: true })
+                        if server_in_tx
+                            .send(ServerInput::AuthResponse { accept: true })
+                            .await
                             .is_err()
                         {
                             break;
@@ -321,8 +271,9 @@ async fn control_loop(
                         // Accept the NLA so the RDP session starts, then collect the second
                         // factor (TOTP / web approval) on a Warpgate-rendered holding screen
                         // before connecting to the target.
-                        if helper_in_tx
-                            .send(ServerHelperInput::AuthResponse { accept: true })
+                        if server_in_tx
+                            .send(ServerInput::AuthResponse { accept: true })
+                            .await
                             .is_err()
                         {
                             break;
@@ -330,8 +281,8 @@ async fn control_loop(
                         match run_hold_screen(
                             &services,
                             &interactive,
-                            &mut frames,
-                            &helper_in_tx,
+                            &mut events,
+                            &server_in_tx,
                             screen,
                         )
                         .await
@@ -349,7 +300,7 @@ async fn control_loop(
                                             connect_backend(
                                                 &services,
                                                 &server_handle,
-                                                &helper_in_tx,
+                                                &server_in_tx,
                                                 user_info,
                                                 target,
                                                 options,
@@ -360,70 +311,45 @@ async fn control_loop(
                                     }
                                     Err(error) => {
                                         warn!(%error, "Authorization failed after second factor");
-                                        let _ = helper_in_tx.send(ServerHelperInput::Shutdown);
+                                        let _ = server_in_tx.send(ServerInput::Shutdown).await;
                                         break;
                                     }
                                 }
                             }
                             Ok(None) => {
                                 warn!("Interactive authentication failed");
-                                let _ = helper_in_tx.send(ServerHelperInput::Shutdown);
+                                let _ = server_in_tx.send(ServerInput::Shutdown).await;
                                 break;
                             }
                             Err(error) => {
                                 warn!(%error, "Holding-screen error");
-                                let _ = helper_in_tx.send(ServerHelperInput::Shutdown);
+                                let _ = server_in_tx.send(ServerInput::Shutdown).await;
                                 break;
                             }
                         }
                     }
                     Ok(DesktopAuthOutcome::Failed) => {
                         warn!("Authentication failed");
-                        let _ =
-                            helper_in_tx.send(ServerHelperInput::AuthResponse { accept: false });
+                        let _ = server_in_tx
+                            .send(ServerInput::AuthResponse { accept: false })
+                            .await;
                     }
                     Err(error) => {
                         warn!(%error, "Authentication error");
-                        let _ =
-                            helper_in_tx.send(ServerHelperInput::AuthResponse { accept: false });
+                        let _ = server_in_tx
+                            .send(ServerInput::AuthResponse { accept: false })
+                            .await;
                     }
                 }
                 continue;
             }
-            ServerHelperEvent::Size { width, height } => {
+            ServerEvent::Size { width, height } => {
                 screen = warpgate_desktop_ui::Screen { width, height };
                 continue;
             }
-            ServerHelperEvent::Error { message } => {
-                warn!(%message, "RDP serve helper reported an error");
-                continue;
-            }
-            ServerHelperEvent::Disconnected => break,
-
             // Viewer input: record it for audit (like native VNC) then forward to the
             // target. `break` once the target client is gone (send fails).
-            ServerHelperEvent::Pointer { x, y, buttons } => DesktopInput::Pointer { x, y, buttons },
-            ServerHelperEvent::Scancode {
-                code,
-                extended,
-                down,
-            } => DesktopInput::Scancode {
-                code,
-                extended,
-                down,
-            },
-            ServerHelperEvent::Key { keysym, down } => DesktopInput::Key { keysym, down },
-            ServerHelperEvent::Wheel {
-                x,
-                y,
-                vertical,
-                delta,
-            } => DesktopInput::Wheel {
-                x,
-                y,
-                vertical,
-                delta,
-            },
+            ServerEvent::Input(input) => input,
         };
 
         // Reached only for the viewer-input variants above.
