--- warpgate-rdp-helper/Cargo.toml
@@ -1,29 +1,0 @@
-[package]
-name = "warpgate-rdp-helper"
-version = "0.23.4"
-edition = "2021"
-license = "Apache-2.0"
-publish = false
-description = "RDP helper for Warpgate, carrying the target-facing client (`connect`) and viewer-facing server (`serve`) as subcommands. Built as a separate binary and embedded into the main executable by warpgate-protocol-rdp."
-
-[dependencies]
-anyhow = "1.0"
-async-trait = "0.1"
-bytes = "1"
-futures = "0.3"
-ironrdp = { version = "0.17", features = ["connector", "session", "graphics", "input"] }
-ironrdp-blocking = "0.10"
-ironrdp-server = "0.13"
-rustls = { version = "0.23", default-features = false, features = ["ring", "std", "tls12", "logging"] }
-rustls-native-certs = "0.8"
-rustls-pemfile = "2"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-sspi = { version = "0.21", features = ["network_client"] }
-tokio = { version = "1", features = ["full"] }
-tokio-util = { version = "0.7", features = ["codec"] }
-tracing = "0.1"
-tracing-subscriber = { version = "0.3", features = ["env-filter"] }
-# Shared IPC message types with warpgate-protocol-rdp.
-warpgate-rdp-ipc = { path = "../warpgate-rdp-ipc" }
-x509-cert = { version = "0.2", default-features = false, features = ["std"] }
